added custom colors

+++ source/types.hpp
@@ -1,5 +1,5 @@
 #include <SDL3/SDL.h>
 
 typedef struct {
-    Uint8 r, g, b, a;
+    float r, g, b, a;
 } Color;
